add Step2

+++ Lesson20/src/main/java/YA/Lesson20/DAO/BucketRepo.java
@@ -3,5 +3,5 @@ package YA.Lesson20.DAO;
 import YA.Lesson20.domain.Bucket;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BucketRepo extends JpaRepository<Bucket,Long> {
+public interface BucketRepo extends JpaRepository<Bucket,Integer> {
 }
