+++ Lesson20/src/main/java/YA/Lesson20/DAO/PeriodicalRepo.java
@@ -3,5 +3,5 @@ package YA.Lesson20.DAO;
 import YA.Lesson20.domain.Periodical;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PeriodicalRepo extends JpaRepository<Periodical,Long> {
+public interface PeriodicalRepo extends JpaRepository<Periodical,Integer> {
 }
